// update_app_01082026/Railmate_App-FIXED-final/Railmate_App-main-FIXED/apk/types/database.types.ts
export type TrainClass =
  | 'SHOVON'
  | 'SHOVON_CHAIR'
  | 'SNIGDHA'
  | 'AC_BERTH'
  | 'AC_SEAT'
  | 'FIRST_BERTH'
  | 'FIRST_SEAT'
  | 'AC_S_CHAIR';